[refactor] Edits added after review
