+++ iOS-FakeNFT-Extended/Services/StorageKeys.swift
@@ -1,0 +1,16 @@
+//
+//  Constants.swift
+//  iOS-FakeNFT-Extended
+//
+//  Created by Ди Di on 12/12/25.
+//
+
+import Foundation
+
+enum StorageKeys {
+        static let name = "profile_name"
+        static let description = "profile_description"
+        static let website = "profile_website"
+        static let photoURL = "profile_photo_url"
+        static let favouriteNFTIds = "favourite_nft_ids"
+}
